move versioning back into OSM
